ml/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0
psycopg2-binary
python-multipart
numpy>=1.24.0
pandas>=2.0.0
rasterio>=1.3.0
Pillow>=10.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
